move prodCard to DETAILS, toggle persistent
