+++ windwatts-ui/src/components/resultPane/details/index.ts
@@ -1,3 +1,4 @@
 export { ProductionDataTable } from "./ProductionDataTable";
+export { ProductionCard } from "./ProductionCard";
 export { WindRose } from "./WindRose";
 export { DetailsTab } from "./DetailsTab";
