+++ windwatts-ui/src/components/resultPane/overview/index.ts
@@ -2,6 +2,5 @@ export { WindSpeedCard } from "./WindSpeedCard";
 export { WindResourceCard } from "./WindResourceCard";
 export { DataSourceLinks } from "./DataSourceLinks";
 export { EnsembleTiles } from "./EnsembleResultsCard";
-export { ProductionCard } from "./ProductionCard";
 export { OverviewTab } from "./OverviewTab";
 export { SummaryCards } from "./SummaryCards";
